clean up/refactor some code,
