+++ ItemVendorLocation/ItemLookup.cs
@@ -3,8 +3,6 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
 using System.Threading.Tasks;
-using Dalamud;
-using Dalamud.Logging;
 using ItemVendorLocation.Models;
 using Lumina.Data.Files;
 using Lumina.Data.Parsing.Layer;
@@ -72,9 +70,6 @@ namespace ItemVendorLocation
         {
             _eventHandlerTypes = Enum.GetValues<EventHandlerType>();
 
-            var stopwatch = new Stopwatch();
-            stopwatch.Start();
-
             _achievements = Service.DataManager.GetExcelSheet<Achievement>();
             _customTalks = Service.DataManager.GetExcelSheet<CustomTalk>();
             _customTalkNestHandlers = Service.DataManager.GetExcelSheet<CustomTalkNestHandlers>();
@@ -105,24 +100,14 @@ namespace ItemVendorLocation
             _gil = _items.GetRow(1);
             _gcSeal = _items.Where(i => i.RowId is >= 20 and <= 22).Select(i => i).ToList();
 
-            stopwatch.Stop();
-            Service.PluginLog.Debug($"{stopwatch.Elapsed}");
-            stopwatch.Start();
             _ = Task.Run(() =>
             {
-                stopwatch.Restart();
                 BuildNpcLocation();
-                stopwatch.Stop();
-                Service.PluginLog.Debug($"BuildNpcLocation: {stopwatch.Elapsed}");
-
-                stopwatch.Restart();
+                ApplyNpcLocationCorrections();
                 BuildVendors();
-                stopwatch.Stop();
-
-                Service.PluginLog.Debug($"BuildVendors {stopwatch.Elapsed}");
                 AddAchievementItem();
+                FixJapaneseShopName();
                 _isDataReady = true;
-                PostProcess();
 #if DEBUG
                 Dictionary<string, uint> noLocationNpcs = new();
                 foreach (KeyValuePair<uint, ItemInfo> items in _itemDataMap)
@@ -142,6 +127,7 @@ namespace ItemVendorLocation
                         noLocationNpcs[npc.Name]++;
                     }
                 }
+
                 Service.PluginLog.Debug("Data is ready");
                 Service.PluginLog.Debug($"Items sold by NPCs with no location: {noLocationNpcs.Values.Aggregate((sum, i) => sum += i)}");
                 Service.PluginLog.Debug("Named NPCs:");
@@ -159,36 +145,6 @@ namespace ItemVendorLocation
 #endif
                 return Task.CompletedTask;
             });
-        }
-
-        // Post processing item map, can be used to fix things
-        private void PostProcess()
-        {
-            // This fix is for non-japanese client
-            // SE is just being lazy on this, hence we have this bug lol
-            if (Service.ClientState.ClientLanguage == ClientLanguage.Japanese)
-            {
-                return;
-            }
-
-            // Look for items that can be purchased from this npc
-            foreach (KeyValuePair<uint, ItemInfo> item in _itemDataMap)
-            {
-                foreach (NpcInfo npcInfo in item.Value.NpcInfos.Where(npcInfo => npcInfo.ShopName == "アイテムの購入"))
-                {
-                    Service.PluginLog.Debug($"{_items.GetRow(item.Key).Name} has ShopName \"アイテムの購入\", correcting to correct one.");
-                    // This correction is for Aenc Ose, who sells "Sheep Equipment Materials", for example.
-                    // A shop is the sub-menu presented at some vendors. Aenc Ose has no such sub-menu, so we simply remove the shop.
-                    npcInfo.ShopName = null;
-                }
-            }
-        }
-
-        // https://discord.com/channels/581875019861328007/653504487352303619/860865002721247261
-        // https://github.com/SapphireServer/Sapphire/blob/a5c15f321f7e795ed7362ae15edaada99ca7d9be/src/world/Manager/EventMgr.cpp#L14
-        private static bool MatchEventHandlerType(uint data, EventHandlerType type)
-        {
-            return ((data >> 16) & (uint)type) == (uint)type;
         }
 
 #if DEBUG
@@ -212,6 +168,7 @@ namespace ItemVendorLocation
                 {
                     continue;
                 }
+
                 BuildVendorInfo(npcBase);
             }
         }
@@ -220,7 +177,7 @@ namespace ItemVendorLocation
         {
             ENpcResident resident = _eNpcResidents.GetRow(npcBase.RowId);
 
-            if (HackyFix_Npc(npcBase, resident))
+            if (FixNpcVendorInfo(npcBase, resident))
             {
                 return;
             }
@@ -238,6 +195,7 @@ namespace ItemVendorLocation
                     SpecialShopCustom specialShopCustom = _specialShops.GetRow(specialShop.Row);
                     AddSpecialItem(specialShopCustom, npcBase, resident);
                 }
+
                 return;
             }
 
@@ -317,7 +275,6 @@ namespace ItemVendorLocation
                                          {
                                              new(scriptArgs[4], _items.GetRow(28).Name.RawString),
                                          },
-
                                          _npcLocations.TryGetValue(npcBase.RowId, out NpcLocation value) ? value : null,
                                          ItemType.SpecialShop);
                         continue;
@@ -386,737 +343,50 @@ namespace ItemVendorLocation
             }
         }
 
-        private void AddSpecialItem(SpecialShopCustom specialShop, ENpcBase npcBase, ENpcResident resident, ItemType type = ItemType.SpecialShop, string shop = null)
-        {
-            if (specialShop == null)
-            {
-                return;
-            }
-
-            foreach (SpecialShopCustom.Entry entry in specialShop.Entries)
-            {
-                if (entry.Result == null || entry.Cost == null)
-                {
-                    continue;
-                }
-
-                foreach (SpecialShopCustom.ResultEntry result in entry.Result)
-                {
-                    if (result.Item.Value == null)
-                    {
-                        continue;
-                    }
-
-                    if (result.Item.Value.Name == string.Empty)
-                    {
-                        continue;
-                    }
-
-                    List<Tuple<uint, string>> costs = (from e in entry.Cost where e.Item != null && e.Item.Value.Name != string.Empty select new Tuple<uint, string>(e.Count, e.Item.Value.Name)).ToList();
-
-                    string achievementDescription = "";
-                    if (type == ItemType.Achievement)
-                    {
-                        achievementDescription = _achievements.Where(i => i.Item.Value == result.Item.Value).Select(i => i.Description).First();
-                    }
-
-                    AddItem_Internal(result.Item.Value.RowId, result.Item.Value.Name, npcBase.RowId, resident.Singular, shop,
-                        costs, _npcLocations.TryGetValue(npcBase.RowId, out NpcLocation value) ? value : null, type, achievementDescription);
-                }
-            }
-        }
-
-        private void AddGilShopItem(GilShop gilShop, ENpcBase npcBase, ENpcResident resident, string shop = null)
-        {
-            if (gilShop == null)
-            {
-                return;
-            }
-
-            for (uint i = 0u; ; i++)
-            {
-                try
-                {
-                    GilShopItem item = _gilShopItems.GetRow(gilShop.RowId, i);
-
-                    if (item?.Item.Value == null)
-                    {
-                        break;
-                    }
-
-                    AddItem_Internal(item.Item.Value.RowId, item.Item.Value.Name, npcBase.RowId, resident.Singular,
-                        shop != null ? $"{shop}\n{gilShop.Name}" : gilShop.Name,
-                        new List<Tuple<uint, string>> { new(item.Item.Value.PriceMid, _gil.Name) },
-                        _npcLocations.TryGetValue(npcBase.RowId, out NpcLocation value) ? value : null, ItemType.GilShop);
-                }
-                catch (Exception)
-                {
-                    break;
-                }
-            }
-        }
-
-        private void AddGcShopItem(GCShop gcId, ENpcBase npcBase, ENpcResident resident)
-        {
-            if (gcId == null)
-            {
-                return;
-            }
-
-            Item seal = _gcSeal.Find(i => i.Description.RawString.EndsWith($"{gcId.GrandCompany.Value.Name.RawString}."));
-            if (seal == null)
-            {
-                return;
-            }
-
-            foreach (GCScripShopCategory category in _gcScripShopCategories.Where(i => i.GrandCompany.Row == gcId.GrandCompany.Row))
-            {
-                for (uint i = 0u; ; i++)
-                {
-                    try
-                    {
-                        GCScripShopItem item = _gcScripShopItems.GetRow(category.RowId, i);
-                        if (item == null)
-                        {
-                            break;
-                        }
-
-                        if (item.SortKey == 0)
-                        {
-                            break;
-                        }
-
-                        AddItem_Internal(item.Item.Value.RowId, item.Item.Value.Name, npcBase.RowId, resident.Singular, null, new List<Tuple<uint, string>> { new(item.CostGCSeals, seal.Name) },
-                            _npcLocations.TryGetValue(npcBase.RowId, out NpcLocation value) ? value : null, ItemType.GcShop);
-                    }
-                    catch (Exception)
-                    {
-                        break;
-                    }
-                }
-            }
-        }
-
-        private void AddInclusionShop(InclusionShop inclusionShop, ENpcBase npcBase, ENpcResident resident)
-        {
-            if (inclusionShop == null)
-            {
-                return;
-            }
-
-            foreach (LazyRow<InclusionShopCategory> category in inclusionShop.Category)
-            {
-                if (category.Value.RowId == 0)
-                {
-                    continue;
-                }
-
-                for (uint i = 0; ; i++)
-                {
-                    try
-                    {
-                        InclusionShopSeriesCustom series = _inclusionShopSeries.GetRow(category.Value.InclusionShopSeries.Row, i);
-                        if (series == null)
-                        {
-                            break;
-                        }
-
-                        SpecialShopCustom specialShop = series.SpecialShopCustoms.Value;
-                        AddSpecialItem(specialShop, npcBase, resident, shop: $"{category.Value.Name}\n{specialShop?.Name}");
-                    }
-                    catch (Exception)
-                    {
-                        break;
-                    }
-                }
-            }
-        }
-
-        private void AddFccShop(FccShop shop, ENpcBase npcBase, ENpcResident resident)
-        {
-            if (shop == null)
-            {
-                return;
-            }
-
-            for (int i = 0; i < shop.Item.Length; i++)
-            {
-                Item item = _items.GetRow(shop.Item[i]);
-                if (item == null || item.Name == string.Empty)
-                {
-                    continue;
-                }
-
-                uint cost = shop.Cost[i];
-
-                AddItem_Internal(item.RowId, item.Name, npcBase.RowId, resident.Singular, null, new List<Tuple<uint, string>> { new(cost, _fccName.Text) },
-                    _npcLocations.TryGetValue(npcBase.RowId, out NpcLocation value) ? value : null, ItemType.FcShop);
-            }
-        }
-
-        private void AddItemsInPrehandler(PreHandler preHandler, ENpcBase npcBase, ENpcResident resident)
-        {
-            if (preHandler == null)
-            {
-                return;
-            }
-
-            uint target = preHandler.Target;
-            if (target == 0)
-            {
-                return;
-            }
-
-            if (MatchEventHandlerType(target, EventHandlerType.GilShop))
-            {
-                GilShop gilShop = _gilShops.GetRow(target);
-                AddGilShopItem(gilShop, npcBase, resident);
-                return;
-            }
-
-            if (MatchEventHandlerType(target, EventHandlerType.SpecialShop))
-            {
-                SpecialShopCustom specialShop = _specialShops.GetRow(target);
-                AddSpecialItem(specialShop, npcBase, resident);
-                return;
-            }
-
-            InclusionShop inclusionShop = _inclusionShops.GetRow(target);
-            AddInclusionShop(inclusionShop, npcBase, resident);
-        }
-
-        private void AddItemsInTopicSelect(TopicSelect topicSelect, ENpcBase npcBase, ENpcResident resident)
-        {
-            if (topicSelect == null)
-            {
-                return;
-            }
-
-            foreach (uint data in topicSelect.Shop)
-            {
-                if (data == 0)
-                {
-                    continue;
-                }
-
-                if (MatchEventHandlerType(data, EventHandlerType.SpecialShop))
-                {
-                    SpecialShopCustom specialShop = _specialShops.GetRow(data);
-                    AddSpecialItem(specialShop, npcBase, resident, shop: topicSelect.Name);
-
-                    continue;
-                }
-
-                if (MatchEventHandlerType(data, EventHandlerType.GilShop))
-                {
-                    GilShop gilShop = _gilShops.GetRow(data);
-                    AddGilShopItem(gilShop, npcBase, resident, shop: topicSelect.Name);
-                    continue;
-                }
-
-                PreHandler preHandler = _preHandlers.GetRow(data);
-                AddItemsInPrehandler(preHandler, npcBase, resident);
-            }
-        }
-
-        private void AddCollectablesShop(CollectablesShop shop, ENpcBase npcBase, ENpcResident resident)
-        {
-            if (shop == null)
-            {
-                return;
-            }
-
-            // skip rows without name
-            if (shop.Name.RawString == string.Empty)
-            {
-                return;
-            }
-
-            for (uint i = 0; i < shop.ShopItems.Length; i++)
-            {
-                uint row = shop.ShopItems[i].Value.RowId;
-
-                // 0 is unspecified, we dont need that
-                if (row == 0)
-                {
-                    continue;
-                }
-
-                // 100 should be enough.. unless SE add more subrows in the future
-                for (uint subRow = 0; subRow < 100; subRow++)
-                {
-                    try
-                    {
-                        CollectablesShopItem exchangeItem = _collectablesShopItems.GetRow(row, subRow);
-                        CollectablesShopRewardItem rewardItem = _collectablesShopRewardItems.GetRow(exchangeItem.CollectablesShopRewardScrip.Row);
-                        CollectablesShopRefine refine = _collectablesShopRefines.GetRow(exchangeItem.CollectablesShopRefine.Row);
-                        // filter out junk data
-                        if (exchangeItem.Item.Value == null || exchangeItem.Item.Row <= 1000)
-                        {
-                            continue;
-                        }
-
-                        // This may be confusing, because some things are incorrectly labeled in Lumina
-                        // CollectableShopRewardScrip is column 8, which is actually the ID in CollectableShopRewardItem
-                        // The reward item is the item that you exchange the collectalbe for.
-
-
-                        AddItem_Internal(rewardItem.Item.Value.RowId, rewardItem.Item.Value.Name.RawString, npcBase.RowId,
-                            resident.Singular.RawString, exchangeItem.CollectablesShopItemGroup?.Value?.Name,
-                            new List<Tuple<uint, string>>() {
-                                new(rewardItem.RewardLow, $"{exchangeItem.Item.Value.Name} min collectability of {refine.LowCollectability}"),
-                                new(rewardItem.RewardMid, $"{exchangeItem.Item.Value.Name} min collectability of {refine.MidCollectability}"),
-                                new(rewardItem.RewardHigh, $"{exchangeItem.Item.Value.Name} min collectability of {refine.HighCollectability}"),
-                            }, /* Will build cost later*/
-                            _npcLocations.TryGetValue(npcBase.RowId, out NpcLocation value) ? value : null,
-                            ItemType.CollectableExchange /*Yes this is special shop*/);
-                    }
-                    catch
-                    {
-                        break;
-                    }
-                }
-            }
-        }
-
-        private void AddQuestReward(QuestClassJobReward questReward, ENpcBase npcBase, ENpcResident resident, List<Tuple<uint, string>> cost = null)
-        {
-            if (questReward == null)
-            {
-                return;
-            }
-
-            if (questReward.ClassJobCategory.Row == 0)
-            {
-                return;
-            }
-
-            if (cost == null)
-            {
-                cost = new List<Tuple<uint, string>>();
-
-                // Build the cost first
-                for (uint i = 0; i < questReward.RequiredItem.Length; i++)
-                {
-                    LazyRow<Item> requireItem = questReward.RequiredItem[i];
-                    if (requireItem.Row == 0)
-                    {
-                        break;
-                    }
-
-                    cost.Add(new Tuple<uint, string>(questReward.RequiredAmount[i], requireItem.Value.Name));
-                }
-            }
-
-            // Add the reward items
-            for (uint i = 0; i < questReward.RewardItem.Length; i++)
-            {
-                LazyRow<Item> rewardItem = questReward.RewardItem[i];
-                if (rewardItem.Row == 0)
-                {
-                    break;
-                }
-
-                AddItem_Internal(rewardItem.Row, rewardItem.Value.Name, npcBase.RowId, resident.Singular.RawString, "",
-                    cost, _npcLocations.TryGetValue(npcBase.RowId, out NpcLocation value) ? value : null,
-                    ItemType.QuestReward);
-            }
-        }
-
-        private void AddQuestRewardCost(QuestClassJobReward questReward, ENpcBase npcBase, List<Tuple<uint, string>> cost)
-        {
-            if (questReward == null || cost == null)
-            {
-                return;
-            }
-
-            if (questReward.ClassJobCategory.Row == 0)
-            {
-                return;
-            }
-
-            for (uint i = 0; i < questReward.RewardItem.Length; i++)
-            {
-                LazyRow<Item> rewardItem = questReward.RewardItem[i];
-                if (rewardItem.Row == 0)
-                {
-                    break;
-                }
-                AddItemCost(rewardItem.Row, npcBase.RowId, cost);
-            }
-        }
-
-        private bool HackyFix_Npc(ENpcBase npcBase, ENpcResident resident)
-        {
-            switch (npcBase.RowId)
-            {
-                case 1043463: // horrendous hoarder
-                    // very ugly code and i dont like it, because se inserted new data between rows in patch 6.5 see here(https://github.com/xivapi/ffxiv-datamining/commit/fd1e8189682d52ee239b9037815a54d54b17a7bc#diff-983b68d9961598b3f0a8cecfc05d0f76f93afd0fd31b6a0cfea188ec12a729a1)
-                    // who knows if they will do it again when they add new stuffs to sanctuary -- nuko
-
-                    RawExcelSheet rawExcel = Service.DataManager.GameData.Excel.GetSheetRaw("custom/007/CtsMjiSpecialShop_00789");
-                    Dictionary<string, string> mjiSpecialShopNames = new();
-
-                    string GetNameFromKey(string key)
-                    {
-                        return mjiSpecialShopNames.TryGetValue(key, out string str) ? str : string.Empty;
-                    }
-
-                    foreach (RowParser parser in rawExcel.GetRowParsers())
-                    {
-                        string key = parser.ReadColumn<string>(0);
-                        string name = parser.ReadColumn<string>(1);
-                        mjiSpecialShopNames[key] = name;
-                    }
-
-                    AddSpecialItem(_specialShops.GetRow(1770601), npcBase, resident, ItemType.SpecialShop,
-                                   $"{GetNameFromKey("TEXT_CTSMJISPECIALSHOP_00789_Q1_000_000")}\n{GetNameFromKey("TEXT_CTSMJISPECIALSHOP_00789_Q2_000_000")}");
-                    AddSpecialItem(_specialShops.GetRow(1770659), npcBase, resident, ItemType.SpecialShop,
-                                   $"{GetNameFromKey("TEXT_CTSMJISPECIALSHOP_00789_Q1_000_000")} \n {GetNameFromKey("TEXT_CTSMJISPECIALSHOP_00789_Q2_000_005")}");
-                    AddSpecialItem(_specialShops.GetRow(1770660), npcBase, resident, ItemType.SpecialShop,
-                                   $"{GetNameFromKey("TEXT_CTSMJISPECIALSHOP_00789_Q1_000_000")}\n{GetNameFromKey("TEXT_CTSMJISPECIALSHOP_00789_Q2_000_010")}");
-                    AddSpecialItem(_specialShops.GetRow(1770602), npcBase, resident, ItemType.SpecialShop, $"{GetNameFromKey("TEXT_CTSMJISPECIALSHOP_00789_Q1_000_005")}");
-                    AddSpecialItem(_specialShops.GetRow(1770603), npcBase, resident, ItemType.SpecialShop, $"{GetNameFromKey("TEXT_CTSMJISPECIALSHOP_00789_Q1_000_010")}");
-                    AddSpecialItem(_specialShops.GetRow(1770723), npcBase, resident, ItemType.SpecialShop, $"{GetNameFromKey("TEXT_CTSMJISPECIALSHOP_00789_Q1_000_025")}");
-                    AddSpecialItem(_specialShops.GetRow(1770734), npcBase, resident, ItemType.SpecialShop, $"{GetNameFromKey("TEXT_CTSMJISPECIALSHOP_00789_Q1_000_030")}");
-                    return true;
-                case 1018655: // disreputable priest
-                    AddSpecialItem(_specialShops.GetRow(1769743), npcBase, resident);
-                    AddSpecialItem(_specialShops.GetRow(1769744), npcBase, resident);
-                    AddSpecialItem(_specialShops.GetRow(1770537), npcBase, resident);
-                    return true;
-                case 1016289: // syndony
-                    AddSpecialItem(_specialShops.GetRow(1769635), npcBase, resident);
-                    return true;
-                case 1025047: // gerolt but in eureka
-                    for (uint i = 1769820; i <= 1769834; i++)
-                    {
-                        SpecialShopCustom specialShop = _specialShops.GetRow(i);
-                        AddSpecialItem(specialShop, npcBase, resident);
-                    }
-                    return true;
-                case 1025763: // doman junkmonger
-                    AddGilShopItem(_gilShops.GetRow(262919), npcBase, resident);
-                    return true;
-                case 1027123: // eureka expedition artisan
-                    AddSpecialItem(_specialShops.GetRow(1769934), npcBase, resident);
-                    AddSpecialItem(_specialShops.GetRow(1769935), npcBase, resident);
-                    return true;
-                case 1027124: // eureka expedition scholar
-                    AddSpecialItem(_specialShops.GetRow(1769937), npcBase, resident);
-                    return true;
-                case 1033921: // faux
-                    AddSpecialItem(_specialShops.GetRow(1770282), npcBase, resident);
-                    return true;
-                case 1027566: // Limbeth, Resplendent Tool Exchange
-                    // we only need the first three npc data (the last one is CustomTalk, we dont need it here)
-                    // the first one is from CollectablesShopItem and the last one is from SpecialShop
-                    AddCollectablesShop(_collectablesShops.GetRow(npcBase.ENpcData[0]), npcBase, resident);
-                    // the second one is obsolete materials, even though they are not craftable, still add them anyway
-                    AddCollectablesShop(_collectablesShops.GetRow(npcBase.ENpcData[1]), npcBase, resident);
-                    // after adding all the items, build the cost
-                    AddSpecialItem(_specialShops.GetRow(npcBase.ENpcData[2]), npcBase, resident);
-                    return true;
-                case 1035014: // Spanner, Skysteel Tool Exchange (but it doesnt seem to do anything???)
-                    // NPCData:
-                    // 0 - Story
-                    // 1 - Default talk
-                    // 2 - CollectableShop
-                    // 3 ~ 5 - PreHandler
-                    AddCollectablesShop(_collectablesShops.GetRow(npcBase.ENpcData[2]), npcBase, resident);
-                    for (int i = 3; i <= 5; i++)
-                    {
-                        PreHandler preHandler = _preHandlers.GetRow(npcBase.ENpcData[i]);
-                        AddItemsInPrehandler(preHandler, npcBase, resident);
-                    }
-                    return true;
-                case 1032900:
-                    // NPCData:
-                    // 0 - Story id
-                    // 1 - SwitchTalk
-                    // 2 ~ 3 SpecialShop 
-                    // 4 - CollectableShop
-                    // 5 - SpecialShop
-                    // 6 - GilShop
-                    // 7 - 8 PreHandler (Replica)
-
-                    AddCollectablesShop(_collectablesShops.GetRow(npcBase.ENpcData[4]), npcBase, resident);
-                    AddSpecialItem(_specialShops.GetRow(npcBase.ENpcData[2]), npcBase, resident);
-                    AddSpecialItem(_specialShops.GetRow(npcBase.ENpcData[3]), npcBase, resident);
-                    AddSpecialItem(_specialShops.GetRow(npcBase.ENpcData[5]), npcBase, resident);
-                    AddGilShopItem(_gilShops.GetRow(npcBase.ENpcData[6]), npcBase, resident);
-                    AddItemsInPrehandler(_preHandlers.GetRow(npcBase.ENpcData[7]), npcBase, resident);
-                    AddItemsInPrehandler(_preHandlers.GetRow(npcBase.ENpcData[8]), npcBase, resident);
-                    return true;
-
-                // add quest rewards, like relic weapons, to item list
-                // but this needs to upadte every time when a new patch drops
-                // hopefully someone can find a better way to handle this -- nuko
-                case 1035012: // Emeny
-                    // 14, 15, 19 -- SkySteel tool
-                    for (uint i = 0; i <= 10; i++)
-                    {
-                        QuestClassJobReward questClassJobReward = _questClassJobRewards.GetRow(14, i);
-                        AddQuestReward(questClassJobReward, npcBase, resident);
-                        questClassJobReward = _questClassJobRewards.GetRow(15, i);
-                        AddQuestReward(questClassJobReward, npcBase, resident);
-                        questClassJobReward = _questClassJobRewards.GetRow(19, i);
-                        AddQuestReward(questClassJobReward, npcBase, resident);
-                    }
-                    return true;
-                case 1016135: // Ardashir
-
-                    List<Tuple<uint, string>> GetCost(uint i)
-                    {
-                        return i switch
-                        {
-                            3 => new List<Tuple<uint, string>>
-                            {
-                                new(1, _items.GetRow(13575).Name), new(1, _items.GetRow(13576).Name),
-                            },
-                            5 => new List<Tuple<uint, string>>
-                            {
-                                new(1, _items.GetRow(13577).Name), new(1, _items.GetRow(13578).Name),new(1, _items.GetRow(13579).Name),new(1, _items.GetRow(13580).Name),
-                            },
-                            6 => new List<Tuple<uint, string>>
-                            {
-                                new(5, _items.GetRow(14899).Name),
-                            },
-                            7 => new List<Tuple<uint, string>>
-                            {
-                                // The amounts are uncertain, so will use the maximum amount
-                                new(60, _items.GetRow(15840).Name),new(60, _items.GetRow(15841).Name),
-                            },
-                            8 => new List<Tuple<uint, string>>
-                            {
-                                new(50, _items.GetRow(16064).Name)
-                            },
-                            9 => new List<Tuple<uint, string>>
-                            {
-                                new(1, _items.GetRow(16932).Name)
-                            },
-                            10 => new List<Tuple<uint, string>>
-                            {
-                                new(1, _items.GetRow(16934).Name)
-                            },
-                            _ => null
-                        };
-                    }
-
-                    // 3 ~ 10 Anima Weapons
-                    for (uint i = 3; i <= 10; i++)
-                    {
-                        for (uint j = 0; j <= 12; j++)
-                        {
-                            QuestClassJobReward questClassJobReward = _questClassJobRewards.GetRow(i, j);
-                            AddQuestReward(questClassJobReward, npcBase, resident);
-                            AddQuestRewardCost(questClassJobReward, npcBase, GetCost(i));
-                        }
-                    }
-                    return true;
-                case 1032903: // gerolt Resistance Weapons
-                    // Build the cost/required items manually, they dont exist in the sheet
-                    for (uint i = 0; i <= 16; i++)
-                    {
-                        QuestClassJobReward questClassJobReward = _questClassJobRewards.GetRow(12, i);
-                        AddQuestReward(questClassJobReward, npcBase, resident, new List<Tuple<uint, string>>
-                        {
-                            new(4, _items.GetRow(30273).Name),
-                        });
-                    }
-                    return true;
-
-                case 1032905: // Zlatan
-                    // Build the cost/required items manually, they dont exist in the sheet
-                    // IL 485
-                    for (uint i = 0; i <= 16; i++)
-                    {
-                        QuestClassJobReward questClassJobReward = _questClassJobRewards.GetRow(13, i);
-                        AddQuestReward(questClassJobReward, npcBase, resident, new List<Tuple<uint, string>>
-                        {
-                            new(4, _items.GetRow(30273).Name),
-                        });
-                    }
-
-                    // build reward items first, then we manually add cost/required items
-                    // code is messy, this could be more optimized and readable, but leave it as it is for now -- nuko
-                    for (uint i = 0; i <= 16; i++)
-                    {
-                        // IL 500
-                        QuestClassJobReward questClassJobReward = _questClassJobRewards.GetRow(17, i);
-                        AddQuestReward(questClassJobReward, npcBase, resident);
-                        AddQuestRewardCost(questClassJobReward, npcBase, new List<Tuple<uint, string>>
-                        {
-                            new(20, _items.GetRow(31573).Name),
-                            new(20, _items.GetRow(31574).Name),
-                            new(20, _items.GetRow(31575).Name),
-                        });
-
-                        // IL 500 #2
-                        questClassJobReward = _questClassJobRewards.GetRow(18, i);
-                        AddQuestReward(questClassJobReward, npcBase, resident);
-                        AddQuestRewardCost(questClassJobReward, npcBase, new List<Tuple<uint, string>>
-                        {
-                            new(6, _items.GetRow(31576).Name)
-                        });
-
-                        // IL 510
-                        questClassJobReward = _questClassJobRewards.GetRow(20, i);
-                        AddQuestReward(questClassJobReward, npcBase, resident);
-                        AddQuestRewardCost(questClassJobReward, npcBase, new List<Tuple<uint, string>>
-                        {
-                            new(15, _items.GetRow(32956).Name)
-                        });
-
-                        // IL 515
-                        questClassJobReward = _questClassJobRewards.GetRow(21, i);
-                        AddQuestReward(questClassJobReward, npcBase, resident);
-                        AddQuestRewardCost(questClassJobReward, npcBase, new List<Tuple<uint, string>>
-                        {
-                            new(15, _items.GetRow(32959).Name)
-                        });
-
-                        // IL 535
-                        questClassJobReward = _questClassJobRewards.GetRow(22, i);
-                        AddQuestReward(questClassJobReward, npcBase, resident);
-                        AddQuestRewardCost(questClassJobReward, npcBase, new List<Tuple<uint, string>>
-                        {
-                            new(15, _items.GetRow(33767).Name)
-                        });
-                    }
-
-                    return true;
-                default:
-                    if (_shbFateShopNpc.TryGetValue(npcBase.RowId, out uint value))
-                    {
-                        AddSpecialItem(_specialShops.GetRow(value), npcBase, resident);
-                        return true;
-                    }
-
-                    return false;
-            }
-        }
-
-        private void AddAchievementItem()
-        {
-            for (uint i = 1006004u; i <= 1006006; i++)
-            {
-                ENpcBase npcBase = _eNpcBases.GetRow(i);
-                ENpcResident resident = _eNpcResidents.GetRow(i);
-
-                for (uint j = 1769898u; j <= 1769906; j++)
-                {
-                    AddSpecialItem(_specialShops.GetRow(j), npcBase, resident, ItemType.Achievement);
-                }
-            }
-        }
-
-        private void AddItem_Internal(uint itemId, string itemName, uint npcId, string npcName, string shopName, List<Tuple<uint, string>> cost, NpcLocation npcLocation, ItemType type,
-            string achievementDesc = "")
-        {
-            if (itemId == 0)
-            {
-                return;
-            }
-
-            if (!_itemDataMap.ContainsKey(itemId))
-            {
-                _itemDataMap.Add(itemId, new ItemInfo
-                {
-                    Id = itemId,
-                    Name = itemName,
-                    NpcInfos = new List<NpcInfo> { new() { Id = npcId, Location = npcLocation, Costs = cost, Name = npcName, ShopName = shopName } },
-                    Type = type,
-                    AchievementDescription = achievementDesc,
-                });
-                return;
-            }
-
-            if (!_itemDataMap.TryGetValue(itemId, out ItemInfo itemInfo))
-            {
-                _ = _itemDataMap.TryAdd(itemId, itemInfo = new ItemInfo
-                {
-                    Id = itemId,
-                    Name = itemName,
-                    NpcInfos = new List<NpcInfo> { new() { Id = npcId, Location = npcLocation, Costs = cost, Name = npcName, ShopName = shopName } },
-                    Type = type,
-                    AchievementDescription = achievementDesc,
-                });
-            }
-
-            if (type == ItemType.Achievement && itemInfo.Type != ItemType.Achievement)
-            {
-                itemInfo.Type = ItemType.Achievement;
-                itemInfo.AchievementDescription = achievementDesc;
-            }
-
-
-            if (itemInfo.NpcInfos.Find(j => j.Id == npcId) == null)
-            {
-                itemInfo.NpcInfos.Add(new NpcInfo { Id = npcId, Location = npcLocation, Name = npcName, Costs = cost, ShopName = shopName });
-            }
-        }
-
-        private void AddItemCost(uint itemId, uint npcId, List<Tuple<uint, string>> cost)
-        {
-            if (itemId == 0)
-            {
-                return;
-            }
-
-            if (!_itemDataMap.TryGetValue(itemId, out ItemInfo itemInfo))
-            {
-                Service.PluginLog.Error($"Failed to get value for ItemId \"{itemId}\" when adding item cost, did you call AddItemCost before the item is added to datamap?");
-                return;
-            }
-
-            NpcInfo result = itemInfo.NpcInfos.Find(i => i.Id == npcId);
-            if (result == null)
-            {
-                Service.PluginLog.Error($"Failed to find npcId \"{npcId}\" for ItemId \"{itemId}\" when adding item cost, did you call AddItemCost before the item is added to datamap?");
-                return;
-            }
-
-            result.Costs.AddRange(cost);
-        }
-
 #if DEBUG
         public void BuildDebugNpcLocation(uint npcId)
         {
-            foreach (TerritoryType sTerritoryType in _territoryType)
+            HashSet<uint> addedAetheryte = new();
+            ExcelSheet<Aetheryte> aetheryteSheet = Service.DataManager.GetExcelSheet<Aetheryte>();
+            foreach (TerritoryType territory in aetheryteSheet!.Where(i => i.Territory.Value != null && i.Territory.Row != 1).Select(i => i.Territory.Value))
             {
-                string bg = sTerritoryType.Bg.ToString();
-                if (string.IsNullOrEmpty(bg))
+                if (addedAetheryte.Contains(territory.RowId))
                 {
                     continue;
                 }
 
-                string lgbFileName = $"bg/{bg[..(bg.IndexOf("/level/", StringComparison.Ordinal) + 1)]}level/planevent.lgb";
-                LgbFile sLgbFile = Service.DataManager.GetFile<LgbFile>(lgbFileName);
-                if (sLgbFile == null)
-                {
-                    continue;
-                }
+                ParseLgbFile(GetLgbFileFromBg(territory.Bg), territory, npcId);
+                addedAetheryte.Add(territory.RowId);
+            }
 
-                ParseLgbFile(sLgbFile, sTerritoryType, npcId);
+            foreach (TerritoryType territory in _territoryType.Where(type =>
+                     {
+                         ContentFinderCondition condition = type.ContentFinderCondition.Value;
+                         // eureka, bozja, gathering
+                         return condition?.ContentType.Row is 26 or 29 or 16;
+                     }))
+            {
+                ParseLgbFile(GetLgbFileFromBg(territory.Bg), territory, npcId);
             }
 
             ExcelSheet<Level> levels = Service.DataManager.GetExcelSheet<Level>();
-            foreach (Level level in levels)
+            foreach (Level level in levels!.Where(i => i.Type == 8 && i.Territory.Value != null))
             {
-                // NPC
-                if (level.Type != 8)
-                {
-                    continue;
-                }
-
                 // NPC Id
                 if (level.Object != npcId)
                 {
                     continue;
                 }
 
-                if (level.Territory.Value == null)
+                ENpcBase npcBase = _eNpcBases.GetRow(level.Object);
+                if (npcBase == null)
+                {
+                    continue;
+                }
+
+                bool match = npcBase.ENpcData.Any(data => _eventHandlerTypes.Any(i => MatchEventHandlerType(data, i)));
+
+                if (!match)
                 {
                     continue;
                 }
@@ -1158,11 +428,17 @@ namespace ItemVendorLocation
                     {
                         continue;
                     }
-#endif
+
+                    if (npcId == null && _npcLocations.ContainsKey(npcRowId))
+                    {
+                        continue;
+                    }
+#else
                     if (_npcLocations.ContainsKey(npcRowId))
                     {
                         continue;
                     }
+#endif
 
                     ENpcBase npcBase = _eNpcBases.GetRow(npcRowId);
                     if (npcBase == null)
@@ -1188,7 +464,7 @@ namespace ItemVendorLocation
                         Map map = _maps.First(i => i.TerritoryType.Row == sTerritoryType.RowId && i.MapIndex == mapId);
                         _npcLocations.Add(npcRowId, new NpcLocation(instanceObject.Transform.Translation.X, instanceObject.Transform.Translation.Z, sTerritoryType, map.RowId));
                     }
-                    catch (InvalidOperationException ex)
+                    catch (InvalidOperationException)
                     {
                         _npcLocations.Add(npcRowId, new NpcLocation(instanceObject.Transform.Translation.X, instanceObject.Transform.Translation.Z, sTerritoryType));
                     }
@@ -1197,17 +473,15 @@ namespace ItemVendorLocation
         }
 
         // https://github.com/ufx/GarlandTools/blob/3b3475bca6f95c800d2454f2c09a3f1eea0a8e4e/Garland.Data/Modules/Territories.cs
+#if DEBUG
+        public void BuildNpcLocation()
+#else
         private void BuildNpcLocation()
+#endif
         {
-            LgbFile GetLgbFileFromBg(string bg)
-            {
-                string lgbFileName = "bg/" + bg[..(bg.IndexOf("/level/", StringComparison.Ordinal) + 1)] + "level/planevent.lgb";
-                return Service.DataManager.GetFile<LgbFile>(lgbFileName);
-            }
-
-            List<uint> addedAetheryte = new();
+            HashSet<uint> addedAetheryte = new();
             ExcelSheet<Aetheryte> aetheryteSheet = Service.DataManager.GetExcelSheet<Aetheryte>();
-            foreach (var territory in aetheryteSheet!.Where(i => i.Territory.Value != null && i.Territory.Row != 1).Select(i => i.Territory.Value))
+            foreach (TerritoryType territory in aetheryteSheet!.Where(i => i.Territory.Value != null && i.Territory.Row != 1).Select(i => i.Territory.Value))
             {
                 if (addedAetheryte.Contains(territory.RowId))
                 {
@@ -1219,11 +493,11 @@ namespace ItemVendorLocation
             }
 
             foreach (TerritoryType territory in _territoryType.Where(type =>
-            {
-                ContentFinderCondition condition = type.ContentFinderCondition.Value;
-                // eureka, bozja, gathering
-                return condition?.ContentType.Row is 26 or 29 or 16;
-            }))
+                     {
+                         ContentFinderCondition condition = type.ContentFinderCondition.Value;
+                         // eureka, bozja, gathering
+                         return condition?.ContentType.Row is 26 or 29 or 16;
+                     }))
             {
                 ParseLgbFile(GetLgbFileFromBg(territory.Bg), territory);
             }
@@ -1245,9 +519,9 @@ namespace ItemVendorLocation
                 bool match = npcBase.ENpcData.Any(data => _eventHandlerTypes.Any(i => MatchEventHandlerType(data, i)));
 
                 if (!match)
+                {
                     continue;
-
-                Service.PluginLog.Debug($"{level.X}, {level.Z}, {level.Territory.Row}, {level.Object}");
+                }
 
                 _npcLocations.Add(level.Object, new NpcLocation(level.X, level.Z, level.Territory.Value));
             }
@@ -1263,12 +537,24 @@ namespace ItemVendorLocation
                     _npcLocations.Add(id.Row, new NpcLocation(0, 0, housingTerrotry));
                 }
             }
-            ManualItemCorrections.ApplyCorrections(_npcLocations);
+        }
+
+        private LgbFile GetLgbFileFromBg(string bg)
+        {
+            string lgbFileName = "bg/" + bg[..(bg.IndexOf("/level/", StringComparison.Ordinal) + 1)] + "level/planevent.lgb";
+            return Service.DataManager.GetFile<LgbFile>(lgbFileName);
         }
 
         public ItemInfo GetItemInfo(uint itemId)
         {
             return !_isDataReady ? null : _itemDataMap.TryGetValue(itemId, out ItemInfo itemInfo) ? itemInfo : null;
+        }
+
+        // https://discord.com/channels/581875019861328007/653504487352303619/860865002721247261
+        // https://github.com/SapphireServer/Sapphire/blob/a5c15f321f7e795ed7362ae15edaada99ca7d9be/src/world/Manager/EventMgr.cpp#L14
+        private static bool MatchEventHandlerType(uint data, EventHandlerType type)
+        {
+            return ((data >> 16) & (uint)type) == (uint)type;
         }
 
         // https://github.com/SapphireServer/Sapphire/blob/a5c15f321f7e795ed7362ae15edaada99ca7d9be/src/world/Event/EventHandler.h#L48-L83
@@ -1278,7 +564,7 @@ namespace ItemVendorLocation
             CustomTalk = 0x000B,
             GcShop = 0x0016,
             SpecialShop = 0x001B,
-            FcShop = 0x002A,    // not sure how these numbers were obtained by the folks at sapphire. This works for my isolated use case though I guess.
+            FcShop = 0x002A, // not sure how these numbers were obtained by the folks at sapphire. This works for my isolated use case though I guess.
             TopicSelect = 0x32,
             PreHandler = 0x36,
             InclusionShop = 0x3a, // 0x38 seems to work too?
